vpr: Fall back to using no bounding box if no path is found
If the router is unable to find a path given the default bounding box,
fallback to using no bounding box.

This ensures that if the only path in the architecture connecting source
and destination is outside the bounding box it will still be found.

Since this failure would usually have resulted in us exiting with an
unroute the run-time overhead of re-trying in those cases seems
reasonable. For the typical case where a path does exist there should be
no run-time overhead.